Revert "Smart 2"
