+++ app/components/screens/favorites/AuthButton.tsx
@@ -1,10 +1,11 @@
-import Link from 'next/link'
 import { FC } from 'react'
+
+import Link from 'next/link'
 
 const AuthButton: FC<{ slug: string }> = ({ slug }) => {
 	return (
 		<Link href={`/auth?redirect=${slug}`}>
-			<a className="btn-primary px-10 py-4">Войти</a>
+			<a className='btn-primary px-10 py-4'>Войти</a>
 		</Link>
 	)
 }
