+++ app/components/screens/favorites/useBookmarks.ts
@@ -1,10 +1,9 @@
-import { useRouter } from 'next/router'
+import { toastError } from '../../../utils/toast-error'
 import { useQuery } from 'react-query'
-
-import { PortalService } from '../../../../api/portal.service'
-import { useAuth } from '../../../../hooks/useAuth'
-import { IPagination } from '../../../../shared/types/movie.types'
-import { toastError } from '../../../../utils/toast-error'
+import { useRouter } from 'next/router'
+import { PortalService } from '../../../api/portal.service'
+import { IPagination } from '../../../shared/types/movie.types'
+import { useAuth } from '../../../hooks/useAuth'
 
 export const useBookmarks = () => {
 	const { user } = useAuth()
@@ -23,10 +22,13 @@ export const useBookmarks = () => {
 				toastError(e, 'get bookmark')
 			},
 			enabled: isUser,
-		}
+		},
 	)
 	const movies = data?.data
 	const pagination = data?.pagination as IPagination
 
+
 	return { isLoading, movies, pagination, isUser }
 }
+
+
