--- app/components/screens/my/menu.data.ts
@@ -1,20 +1,0 @@
-import { LINKS } from '../../../config/links'
-import { IMenuItem } from '../../layout/Navigation/MenuContainer/menu.interface'
-
-export const myMenu: IMenuItem[] = [
-	{
-		icon: 'MdBookmarkBorder',
-		link: LINKS.FAVORITES,
-		title: 'Избранное',
-	},
-	{
-		icon: 'MdBookmarkBorder',
-		link: LINKS.HISTORY,
-		title: 'История просмотра',
-	},
-	{
-		icon: 'MdBookmarkBorder',
-		link: LINKS.RATING,
-		title: 'Оценки',
-	},
-]
