--- app/components/screens/my/rating/rating.tsx
@@ -1,7 +1,0 @@
-import { FC } from 'react'
-
-const Rating: FC = () => {
-	return <div>Rating</div>
-}
-
-export default Rating
