+++ app/config/links.ts
@@ -7,18 +7,14 @@ export const LINKS = {
 	REGISTER: '/register', //регистрация
 	PROFILE: '/profile',//профиль
 	RUBLE: '/settings/subscriptions',//купить за рубль
-	FAVORITES: '/my/favorites', //избранное
-	HISTORY: '/my/history', //История просмотра
-	RATING:'/my/rating', // мои оценки
+	FAVORITES: '/favorites', //избранное
+	HISTORY: '/history', //История просмотра
 	SEARCH: '/search', //поиск с фильтрами
 	SETTINGS: '/settings', //страница настроек
 	SUBSCRIPTIONS: '/settings/subscriptions', //подписки
 	PROMOCODE: '/settings/promocode', //промокод
 	DEVICES: '/settings/devices', //устройства
 	ACCOUNT: '/settings/account', //аккаунт
-	FILMS:'/genres/100',
-	SERIALS:'/genres/39',
-	MULTFILMS:'/genres/20',
 
 
 	 FEEDBACK:'/feedback'//Сообщить об ошибке
